Simplify quiz logic
